api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pillow==11.0.0
numpy==2.1.0
scikit-learn==1.5.2
scikit-image==0.24.0
colour-science==0.4.4
python-multipart==0.0.12
